python/pyproject.toml
[project]
name = "mcp-k8s-operator-python"
version = "0.1.0"
description = "Python/Kopf prototype for a Kubernetes MCP operator"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "fastmcp>=2.13.0",
  "kopf>=1.37.0",
  "kubernetes>=29.0.0",
  "pyyaml>=6.0.0",
  "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
test = [
  "httpx>=0.27.0",
  "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
